+++ src/model/BrandName.ts
@@ -1,1 +1,15 @@
-export type BrandName = 'Anchor' | 'Appletons' | 'Cosmo' | 'DMC' | 'J&P Coats' | 'Kreinik' | 'Madeira' | 'Presenica' | 'Sullivans';
+/**
+ * An enum representing the various supported floss brand names
+ */
+export enum BrandName {
+    ANCHOR = 'Anchor',
+    APPLETONS = 'Appletons',
+    COSMO = 'Cosmo',
+    DMC = 'DMC',
+    JP_COATS = 'J&P Coats',
+    KREINIK = 'Kreinik',
+    MADEIRA = 'Madeira',
+    PRESENICA = 'Presenica',
+    SULLIVANS = 'Sullivans',
+    UNBRANDED = 'Unbranded'
+}