+++ src/model/BrandedFloss.ts
@@ -1,8 +1,0 @@
-import { BrandName } from './BrandName';
-
-export interface BrandedFloss {
-    brandName: BrandName;
-    colorCode: string;
-    colorName: string;
-    otherBrandEquivalents?: BrandedFloss[]
-}